Adds :void to setUp methods in unit tests examples
As of PHP 7 we can easily add void RTH which is required by PHPUnit 8+.
As we support PHPunit 8+ in laminas-test and PHP 5.6 is deprecated
there is no point of keeping these examples compatible with unsupported
PHP version.

Changed visibility to protected, as it should be the same as parent.

Fixes #15

Signed-off-by: Michał Bundyra <contact@webimpress.com>
Signed-off-by: AdeMS <ademokoginta@gmail.com>